type integer for sign warning
git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@18761 0fceea05-a30d-0410-8a8b-80ef821fb0a1

+++ BDSKMainTableView.m
@@ -202,7 +202,7 @@ enum {
                 }
             } else {
                 // Don't try to go past the new # of rows
-                selectedRow = MIN(selectedRow - 1, newNumberOfRows - 1);
+                selectedRow = MIN((NSUInteger)selectedRow - 1, newNumberOfRows - 1);
                 
                 // Skip all unselectable rows if the delegate responds to -tableView:shouldSelectRow:
                 if ([[self delegate] respondsToSelector:@selector(tableView:shouldSelectRow:)]) {
